Completed day 8
Today I learned that you can select all & copy a document before it loads completely, which results in only part of the document being copied. I was only using ~6000 out of 10000 lines of input, so I was very delayed with solving the problem. I was convinced that I was interpreting part of the question wrong.