Completed day 26
Another first place finish, but with yet another terrible solution. This is my slowest solution yet, but since it was faster to code and execute this than it was to code a better solution, this is what I used.
I plan to come back to this problem later and solve it in a far better way.