Completed day 27
I should have gone straight for the naïve solution today. Instead, I wasted some time trying to find the better solution, because I thought (particularly after yesterday) that *this time* a smart solution would be required. Unfortunately, that was not the case.
I implemented the efficient solution using the Extended Euclidean algorithm anyway, because it was good to learn.