Completed day 29
I messed up today, but wasn't punished too badly.
I started by trying to find a combinatorial solution, because I thought bruteforcing would take too long. When I inevitably realised that that was a terrible idea, I made a bruteforce solution, but it was awful, taking over 20 minutes to run.
That got me over the line, then I decided to figure out why my bruteforcer was so much worse than others'. I changed a few things, and was able to get the runtime down to 10s.

I also had a look at the true, mathematical solution, and was a bit disappointed to see how complicated it is. There was no chance I was ever finding a solution remotely close to the optimal one.