Completed day 30, and won the competition!
Today was the final day, and although there was a slight bug with half of the inputs (mine included), I still came first overall in the competition! It was a lot of fun, I'd definitely do something like this again, though I doubt it'll be quite as easy to win next time.

I've actually been waiting for a problem like today's. The Advent of Code uses a lot of pathfinding problems, so I was ready for one. Sure enough, I coded a solution really fast, but due to that input bug I was a bit delayed submitting a "correct" answer.